Restructured component file structure for better scalablity.
