+++ src/components/DesktopNavMenu/index.jsx
@@ -1,0 +1,7 @@
+import React from "react";
+
+function DesktopNavMenu() {
+  return <div>DesktopNavMenu</div>;
+}
+
+export default DesktopNavMenu;
