+++ src/components/Header/index.jsx
@@ -1,5 +1,5 @@
 import React from "react";
-import Navigation from "./Navigation";
+import Navigation from "../Navigation";
 import { Link } from "react-router-dom";
 
 function Header() {
